move reset reason to system task
