+++ stm32-modules/include/flex-stacker/firmware/system_policy.hpp
@@ -23,4 +23,5 @@ class SystemPolicy {
         -> errors::ErrorCode;
     auto get_serial_number() -> std::array<char, SYSTEM_SERIAL_NUMBER_LENGTH>;
     auto get_door_closed() -> bool;
+    auto last_reset_reason() const -> uint16_t;
 };
