+++ stm32-modules/include/flex-stacker/flex-stacker/system_task.hpp
@@ -87,6 +87,17 @@ class SystemTask {
     }
 
     template <SystemExecutionPolicy Policy>
+    auto visit_message(const messages::GetResetReasonMessage& msg,
+                       Policy& policy) -> void {
+        auto reason = policy.last_reset_reason();
+
+        auto response = messages::GetResetReasonResponse{
+            .responding_to_id = msg.id, .reason = reason};
+        static_cast<void>(
+            _task_registry->send(response, Queues::HostCommsAddress));
+    }
+
+    template <SystemExecutionPolicy Policy>
     auto visit_message(const messages::SetSerialNumberMessage& message,
                        Policy& policy) {
         auto response =
